models refactoring
