+++ models/config.go
@@ -1,0 +1,7 @@
+package models
+
+type Config struct {
+	TargetPort int
+	Hostname   string
+	LogLevel   string
+}
